[runtime] Install headers and library for the private part
We need to make the library and plugin interface available to support
more platforms. On a Plasma Wayland session it's required that KWin
becomes KGlobalAccel. KWin is the only process which can filter key
events and for security reasons KWin will not delegate them to other
applications.

This means KWin needs to create KGlobalAccelD before KWin or any other
process tries to register a global shortcuts triggering kglobalaccel5
to be started.

In addition KWin needs to provide it's own plugin which it can interact
with.

CHANGELOG: private runtime API gets installed to allow KWin to provide plugin for Wayland.
REVIEW: 124186