Introduce an env variable to overwrite the platform name
With the env variable KGLOBALACCELD_PLATFORM the platform name can
be specified, which is being used for searching for the plugin to
load. This is required because some plugins are more specific than
the supported platforms. If the environment variable is not
specified, the QGuiApplication::platformName() is used as before.

CCBUG: 349911
REVIEW: 124266
CHANGELOG: New env variable KGLOBALACCELD_PLATFORM to specify the platform plugin to load.