fix handling of irregular 1st level shifts
if a key has a different symbol on the first level
shift (ie. not upcase, affects non-alphabetic keys)
it's not sufficient to just strip the SHIFT modifier
but the actual symbol for the 1st level shift needs
to be fetched.

Thanks to Yuxuan Shui for initially working on this.

BUG: 341959
REVIEW: 124893