[runtime] Introduce a KGLOBALACCEL_TEST_MODE env variable
Summary:
The idea behind the env variable is to put kgloabalacceld into test mode
and let it operate on a in-memory KConfig instead of the normal rc file.

From auto tests in KWin we know that the
QStandardPath::setTestModeEnabled is not sufficient. If there are
multiple tests interacting with kglobalaccel they interact with each
other in ways we normally don't want.

With this env variable we could get KWin to have an in-memory kconfig
and support better testing of the global shortcuts. We can adjust the
shortcuts without having side effects on other tests.

Reviewers: #frameworks

Differential Revision: https://phabricator.kde.org/D3689