Revert "KGlobalAccel: port to KKeyServer's new method symXModXToKeyQt, to fix numpad keys"
This reverts commit 734fb2eb9763b0e665721cffc3b1a3462b99a715.

I'm reverting as we don't have a fix yet and the next release is too
close to risk any changes. As explained in the original review request
for this feature: touching this code base is dangerous. The code is
extremely fragile.

A month of living in master was not enough to spot the severe regressions
we had. As we are now only a few days away from next release we cannot
expect to discover regressions any fix would cause. I don't really see a
solution to this problem: any change might cause issues especially in
KWin, KScreenLocker and KGlobalAccel. All of them have the problem that
the maintainer of those components is not running X11 and won't spot the
regressions. Futhermore we don't have any test cases for the complete key
handing stack on X11. We won't find the regressions. My personal
recommendation would be to no longer change the X11 side and instead all
together work on Wayland and get it ready. There the world looks better:
we have a sane input stack which is completly unit tested. The first
regression caused by the changes here was discovered by a KWin test case
for Wayland. That's the world where we can change and improve the stack.
Let X die, long live Wayland!

If someone wants to take on the task to improve the X11 stack here I
expect some serious work on the testability of this fragile stack. If we
see that the shortcuts will still match I would be way more confident to
accept a change.

CCMAIL: kde-frameworks-devel@kde.org
BUG: 384597