kglobalacceld: Avoid loading an icon loader for no reason
Summary:
We know this is a QGuiApplication app, so we know it doesn't have any
widgets we also know it doesn't use QtQuick anywhere.

Therefore it won't have any windows and setting a default window icon is
pointless.
This has a surprisingly big overhead as it instantiates the KDE icon theme
engine. Saves near 5% of startup time.

Test Plan: Compiles

Reviewers: apol

Reviewed By: apol

Subscribers: apol, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9047