Add migration helper to migrate dynamic shortcuts to desktop file based shortcuts
Desktop file based shortcuts are saved differently in the config, and when switching something to it we want to keep the user's configuration.

Doing the migration client side is hard because there is no natural place to execute procedural migration code for declarative desktop file shortcuts.

This introduces a new key for shortcut desktop files in /usr/share/kglobalaccel, X-KDE-Migrate-Shortcut. It contains, comma separated, a component and action name from which the old shortcut is read and migrated to the new storage