Fix losing shortcuts when updating the system
Amends 995052f85acd4cf089c4394aa3901c729c0e4f8b.

When pacman updates konsole package, it will delete the desktop file and
create it again.

If the global shortcuts registry is refreshed right after deleting the
desktop file but before a new one is created, it will remove the
corresponding shortcuts component.

When the desktop file is created again, the GlobalShortcutsRegistry will
read the shortcuts from the desktop file and the previously configured
shortcuts will be lost.

This change makes detectAppsWithShortcuts() check the config file if there
are any shortcuts for the app to prevent losing shortcuts.

However, it'd be nice if there were a proper hook or something to refresh
shortcuts after a system update is performed.