registry: do not refresh invalid registries
previously this would result in failed assertions when trying to
construct a component's dbus path by prefixing it with the registry's
dbus path. the latter being empty meant the path ended up like
`component/foo` which is not a valid dbus path (missing leading /).
this would then fail the assertion in
`GlobalShortcutsRegistry::registerComponent` and crash kwin

more ideally we probably shouldn't have a registry if the init fails,
that's a large refactor though