Fix unneeded rewrites of kglobalshortcutrc on startup
**Origin of the bug**

In 2021 in git commit 0e26ec40b7aa19 while porting from int
to QKeySequence following change was done:

```
- static QList<int> keysFromString(const QString &str)
+ static QList<QKeySequence> keysFromString(const QString &str)
{
-    QList<int> ret;
+    QList<QKeySequence> ret;
    if (str == QLatin1String("none")) {
        return ret;
    }
    const QStringList strList = str.split('\t');
    for (const QString &s : strList) {
-        int key = QKeySequence(s)[0];
-        if (key != -1) { // sanity check just in case
+        QKeySequence key = QKeySequence(s);
+        if (!key.isEmpty()) { // sanity check just in case
            ret.append(key);
        }
    }
    return ret;
}
```

The conditions `(key[0] != -1)` and `(!key.isEmpty())` are not
identical but lead to very different results.

The -1 condition is still used in code like
[kglobalaccel.cpp](https://invent.kde.org/frameworks/kglobalaccel/-/blob/v6.19.0/src/kglobalaccel.cpp?ref_type=tags#L597)
like this:

```
if (sequence[i].toCombined() == -1) {
    qWarning() << "Encountered garbage keycode (keycode = -1) in input, not doing anything.";
    return true;
}
```

The QKeySequence().isEmpty() condition in contrast also applies to
an empty sequence means QKeySequence("").isEmpty()==true. An empty
sequence is actually (QKeySequence("")[0].toCombined()>=0)==true
since QKeySequence("")[0].toCombined()==Qt::Key_unknown and
Qt::Key_unknown==0x01ffffff.

**The problem resulting from the bug**

Now within kglobalaccel and kglobalacceld an empty, means not set,
shortcut is internally represented as QList(QKeySequence("")).

This is then written into the kglobalshortcutrc.

But on restoring that empty keysequence from the kglobalshortcutrc it
turned into QList() because of the QKeySequence("").isEmpty()==true
bug described above.

That had now the effect that in this code:

```
QList<QKeySequence> KGlobalAccelD::setShortcutKeys(...)
{
    [...]
    if (isDefault) {
        if (shortcut->defaultKeys() != keys) {
            shortcut->setDefaultKeys(keys);
            scheduleWriteSettings();
        }
        return keys; // doesn't matter
    }
    [...]
}
```

For every empty global shortcut the `(shortcut->defaultKeys() != keys)`
condition was always true because `(QList() != QList(QKeySequence(""))`
is true.

That means after kglobalacceld did restore the shortcuts from its
kglobalshortcutrc every empty shortcut, that was still empty when
KGlobalAccelD::setShortcutKeys was called, actually resulted in
QList() being replaced with QList(QKeySequence("")) AND a
scheduleWriteSettings() being called because the shortcut changed.

That happened for ALL empty shortcuts. Since KGlobalAccelD::setShortcutKeys
is called during the whole startup we had calls to scheduleWriteSettings()
during the whole startup.

The scheduleWriteSettings() actually uses a timer to delay writing
the kglobalshortcutrc by 500ms and compress all scheduleWriteSettings()
calls done within that 500ms into one write operation.

Still depending on how long the startup takes on your system this
resulted then in >=1 rewrites of the kglobalshortcutrc to disk each
500ms during KDE startup.

On my system, where the startuip takes <2 seconds, the kglobalshortcutrc
got written 3 times. Checking the content of the previous kglobalshortcutrc
and the new written kglobalshortcutrc shows that the content is identical
and the write completly unneeded.

**The fix for the bug**

This MR fixes that and makes sure the kglobalshortcutrc is not rewritten
for no reason at every KDE startup by removing the wrong isEmpty()
condition.

An alternate fix could be to change it from `(!key.isEmpty())` to
`(key[0].toCombined() != -1)` what equals exact the logic before the port
and equals to how other such -1 verify code was ported.

I decided to just remove it because 1) to sanity check what we write
ourself makes only sense if you assume that someone is manipulating
the kglobalshortcutrc manually what sounds not like a use-case
we need to guard against and 2) I even think that the assumption
QKeyCombination::toCombined() can return -1 is already a wrong assumption
and a left-over of a 1:1 port from int to QKeySequence.